requirements.txt
numpy==1.21.5
pulsectl==22.3.2
pyserial==3.5
PySimpleGUI==4.60.3
